refactor(AYC-000): move department list ownership to frontend
