+++ ayunis-core-frontend/src/pages/auth/register/api/useRegister.ts
@@ -8,7 +8,6 @@ import {
   computeDepartment,
   refineDepartmentOther,
 } from '@/shared/lib/compute-department';
-import type { RegisterDtoDepartment } from '@/shared/api/generated/ayunisCoreAPI.schemas';
 import { DEPARTMENT_OTHER_MAX_LENGTH } from '@/shared/constants/department';
 import { useTranslation } from 'react-i18next';
 import * as z from 'zod';
@@ -66,7 +65,7 @@ export function useRegister() {
           department: computeDepartment(
             values.department,
             values.departmentOther,
-          ) as RegisterDtoDepartment | undefined,
+          ),
           marketingAcceptance: values.marketingAcceptance,
         },
       },
