+++ ayunis-core-frontend/src/shared/SUMMARY.md
@@ -4,4 +4,4 @@ Foundation layer with UI components, generated API client, and utilities.
 
 The shared layer provides framework-level primitives: 37 shadcn/ui components, an Orval-generated TypeScript API client with TanStack Query hooks (~13,600 lines), Axios HTTP client configuration, i18n locale files for German and English, chat context provider, error boundary components, and utility libraries.
 
-The `api/` directory contains the Orval-generated client in `generated/ayunisCoreAPI.ts` (~10,700 lines of typed React Query hooks for every backend endpoint) and `generated/ayunisCoreAPI.schemas.ts` (~2,900 lines of TypeScript interfaces matching backend DTOs). The `client.ts` file configures the shared Axios instance with base URL and authentication interceptors, while `extract-error-data.ts` provides error response parsing. The `ui/shadcn/` directory houses 37 component primitives (button, dialog, form, table, select, sidebar, card, tabs, etc.) built on Radix UI. The `ui/error-boundary/` provides React error boundary with fallback UI. The `ui/help-link/` provides a HelpLink component for linking to external help center documentation with icon and default variants. The `contexts/chat/` implements a chat context provider for sharing conversation state across components. The `hooks/` directory includes shadcn's `use-mobile` responsive hook. The `locales/` directory contains German and English translation JSON files organized by feature area. The `constants/` directory holds shared constant modules: `department.ts` derives selectable department keys from the backend OpenAPI schema (via the generated `RegisterDtoDepartment` enum) and exports `DEPARTMENT_KEYS`, `DepartmentKey`, and `DEPARTMENT_OTHER_MAX_LENGTH`; `secret-mask.ts` provides secret masking utilities. The `lib/` directory provides DOM patch utilities, shadcn's `cn()` class merging helper, a `help-center.ts` utility for building locale-aware help center URLs, a `computeUsagePercent` helper for calculating clamped percentage values, and `compute-department.ts` which exports `computeDepartment` (resolves department + free-text "other" into the `other:<text>` wire format) and `refineDepartmentOther` (shared Zod superRefine callback for the other-text required check). The `ui/department-field/` directory provides a reusable `DepartmentField` form component (select dropdown with conditional free-text input for "other"). The `config/` directory exports environment-based application configuration. The `assets/brand/` directory contains branding resources.
+The `api/` directory contains the Orval-generated client in `generated/ayunisCoreAPI.ts` (~10,700 lines of typed React Query hooks for every backend endpoint) and `generated/ayunisCoreAPI.schemas.ts` (~2,900 lines of TypeScript interfaces matching backend DTOs). The `client.ts` file configures the shared Axios instance with base URL and authentication interceptors, while `extract-error-data.ts` provides error response parsing. The `ui/shadcn/` directory houses 37 component primitives (button, dialog, form, table, select, sidebar, card, tabs, etc.) built on Radix UI. The `ui/error-boundary/` provides React error boundary with fallback UI. The `ui/help-link/` provides a HelpLink component for linking to external help center documentation with icon and default variants. The `contexts/chat/` implements a chat context provider for sharing conversation state across components. The `hooks/` directory includes shadcn's `use-mobile` responsive hook. The `locales/` directory contains German and English translation JSON files organized by feature area. The `constants/` directory holds shared constant modules: `department.ts` defines the selectable department keys (the frontend owns this list; the backend accepts any string) and exports `DEPARTMENT_KEYS`, `DepartmentKey`, and `DEPARTMENT_OTHER_MAX_LENGTH`; `secret-mask.ts` provides secret masking utilities. The `lib/` directory provides DOM patch utilities, shadcn's `cn()` class merging helper, a `help-center.ts` utility for building locale-aware help center URLs, a `computeUsagePercent` helper for calculating clamped percentage values, and `compute-department.ts` which exports `computeDepartment` (resolves department + free-text "other" into the `other:<text>` wire format) and `refineDepartmentOther` (shared Zod superRefine callback for the other-text required check). The `ui/department-field/` directory provides a reusable `DepartmentField` form component (select dropdown with conditional free-text input for "other"). The `config/` directory exports environment-based application configuration. The `assets/brand/` directory contains branding resources.
